app/requirements.txt
chromadb
Flask
Flask-Session
pycryptodome
psycopg2-binary
Flask-SQLAlchemy
pymongo